Contains Malone and the attack of the killer hash codes: Fix for 425: Consider using Contains method if available
When checking whether or not an entity exists in a navigation property the code was always doing a linear search through the entities. In a lot of cases calling Contains on the collection will be faster. However, Contains may not work correctly in cases where Equals and GetHashCode are overridden. (This is arguably an issue in the app code, but we are trying to be resilient to this.)

The fix is to use Contains either or both of two conditions hold:
1. The navigation property is a HashSet that uses our reference equality comparer. This is what we create internally for null nav properties, but it can also be done explicitly by the app.
2. The entity type being added does not override Equals or GetHashCode. It is still possible that the collection is doing non-standard comparisons without using Equals or GetHashCode, but this seems like a good balance.