CodePlex #677: TPT to TPH - can not use the same name properties in different hierarchies.
Scenario:

A - table A
B : A - table B
B1 : B - table B
C : A - table C
C1 : C - table C
Y - table Y

B (Dependent) -> Y (Principal)
C (Dependent) -> Y (Principal)

Issue 1. The selector used in EntityMappingTransformer.GetDependentColumns
for destinationColumns returns two results because the Name and the
UnpreferredUniqueName are both matched, and results in exception thrown by
Single(). This happens because of the way the database model is
constructed from the conceptual model. Changed the selector to first
select by Name and if not found to select by UnpreferredUniqueName. This
solves the exception reported, but the model still doesn't load because of another exception that complains about the Discriminator condition.

Issue 2. The code adds the discriminator column	to the base type and the	condition properties to	the derived types. In ConfigureDefaultDiscriminator	the discriminator condition is removed if the entity type is not sharing the table with its base type. That results in incorrectly removing the discriminator condition for types B and C because they are not sharing the table with A, even though they are sharing tables with B1 respectively C1. Removed the check of isSharingTableWithBase because code in ModelConfiguration.RemoveRedundantColumnConditions is supposed to remove the discriminators when there is only one table using it. However that code calls StorageMappingFragmentExtensions.RemoveDefaultDiscriminatorCondition which did not seem to work.

Issue 3. Intent of code in StorageMappingFragmentExtensions.RemoveDefaultDiscriminatorCondition is to remove the single condition but the implementation fails to achieve that because of how the key is formed for the internal dictionary. To fix this I replaced that code with a call to ClearConditions.