Migrations: Fix for #1251 - CF Stored Proc :: Invalid migration created when mapping entity with inheritance
Problem: When performing sproc body SQL gen, we were incorrectly dealing with discriminators by lifting the corresponding DbConstant expressions into parameters. However, for sproc bodies the discriminator values need to remain inlined as they are never passed as parameters by the runtime - They are effectively hard-coded because of the one-sproc-per-entity-type rule.
Solution: For sprocs we don't actually need to do parameter lifting in SQL gen because we do the lifting earlier, as part of the dynamic-to-function command tree conversion process. I disabled the lifting on the insert gen code-path as that's the only one relevant for discriminators.

Work Item: 1251