Code First: Fix for #1343 - Stored Procs :: mapping many-many self reference to stored procedures throws if one overrides parameter names
Problem: When configuring parameter names for many-to-many self-refs, the PropertyPath alone is not enough to correctly select which parameter should be configured.
Solution: Introduce a ParameterKey class that contains both a PropertyPath _and_ a bool used to indicate whether we are addressing the left vs. right key when dealing with a many-to-many self-ref.

Work Item: 1343