Knowing Key, Knowing You: A-Ha! (Additional fix for 1192: Making initializers and Migrations play nice)
The problem was that when the MigrationHistory table is created from Migrations it will use the ContextKey specified in the Migrations configuration. If model compatibility is then going to be checked it must also use this context key.

The solution is to moved Migrations configuration discovery to InternalContext and then use the ContextKey discovered when doing compatibility checks.