Not so complex after all... (Fix for 1278: DetectChanges throws when changing complex property instance on change tracking proxy)
This was a regression introduced post alpha 3 by a change to optimize original value tracking in the state manager. The problem is that when a complex object instance is changed in a change tracking proxy the change is updated immediately in the state manager. However, this was not taken into account by DetectChanges which then detected the snapshot difference and tried to make the same change again.

It might be possible to have DetectChanges not process complex object instance changes when change tracking proxies are being used, but this is not straightforward code and I am not confident that it would a) work b) have any significant benefit and c) not introduce more regressions. Therefore, the simple fix here is to make the new optimized code behave in the same way as it always has--that is, allow DetectChanges to make the change even though it has already been recorded.