Dude, where's your key? (Fix for 663: Exception when calling DetectChanges twice with nullable keys)
The issue was due to attempting to store a null FK incorrectly. The code was assuming that it was possible to create a new EntityKey using the PK value(s). But the PK key values are null in this case, so the FK is also null. EntityKey doesn't support null values and instead expects the FK to be stored as null. I checked the code for other places where the same thing might be happening and couldn't find anything.