Migrations: Fix for #1242 - Misleading exception thrown when trying to map entity with composite key, being a dependent of 1-Many relationship
Problem: When generating update trees for dependent entities of IAs whose principal end is required, we first need to change the relationship state to deleted before setting it to added.
Solution: This is actually the same requirement that was previously identified and fixed for one-to-ones (#1190). So, the fix was to broaden the solution to handle all IAs having a required principal.

Work Item: 1242