Dude, that's not my key! (Additional fix for 1192)
The previous checkin for this issue made InternalContext use the context key from the discovered migrations configuration if possible. This worked for compatibility checking in initializers, but caused a problem with scaffolding initial create. The problem was that when scaffolding an initial create the first step involves creating a configuration. Later this is then discovered and the context key is used from it instead of using the default. The fix is to have initial create always use the default.