Six herrings or half a dozen mackerel (Fix for 906: Database.UseTransaction may perform context initialization)
This is a bit of a red herring given where we landed with transactions and initializers. First, in order to use UseTransaction you need to have a connection to a database, so the initializer here cannot be being used to create the database. Second, initialization always uses a cloned connection because attempting to reuse the same connection caused more problems than it fixed (see 771). Since the initializer is always using a cloned connection it means that it will never be using the transaction associated with the current connection.

Therefore, regardless of whether the code is changed the initializer will not make use of a transaction supplied by UseTransaction. However, given that there is no need to initialize the database on a call to UseTransaction and this is a simple change it has been made.