Issue 1202: Properties configuration .HasPrecision(byte) doesn't affect DateTime properties
When changing the column type we were copying the facets from the previous type if they still applied. This means that when changing from datetime to datetime2 the precision remained a constant facet.

Solution:
When changing the column type only carry over the values of the facets from the previous type and only if they differ from the default.

Workitem: 1202