Just say NO (to initialization)... (Fix for 291 and 1316: Disable and re-enable database initialization correctly)
DbContextInfo should disable initialization for the context it is being used with, but this should not disable initialization for later uses of the context in the same app domain outside of DbContextInfo.

The two "obvious" ways of disabling initialization don't work well:
- Calling Database.SetInitializer(null) (or equivalent) has two problems:
  - If use code (say in the context constructor) calls Database.SetInitializer again, then the initializer will be re-enabled.
  - Initialization stays disabled even after this use by DbContextInfo has finished.
- Setting a flag on the constructed context doesn't work because initialization can be triggered in the context constructor before there is a chance to set the flag.

Therefore, the fix here is to add a separate lookup for suppressed initializers that works outside of the normal initialization mechanisms and can be controlled by DbContextInfo or anything else that needs to temporarily suppress initialization.