LeanQueryManufacturing - fix for work item 1390
Queries that produce JOIN statements and contain a call to a limiting operation without first using OrderBy() are producing more complicated SQL than in previous versions. During the generation of the command tree, we visit the DbLimitExpression produced by the call to the limiting operation and, depending on the type of node that is being processed when the DbLimitExpression is encountered, either we nest the limit operation to the child node, or we apply the limit operation to the current node. The code comment indicates that the only reason we would like to have the limit operation nested is in case the current node is a Projection of a Skip operation, or a Projection of a Sort operation. However, the nesting is currently being applied to all projections. In the case of this bug, we are nesting the limit operation because the node that Linq produced for our query is a Projection of a Filter operation (the filter being the WHERE statement in the Linq query).
There are only three cases where a DbLimitExpression will be generated when calling any of the five limiting operations (First(), FirstOrDefault(), Single(), SingleOrDefault(), or Take()):
1.Calling it directly on the IQueryable.
2.Calling it after an OrderBy() on the IQueryable.
3.Calling it after a Skip() on the IQueryable.
The proposed fix is to only push the limit operation below the projection when there is an explicit call to OrderBy() or Skip() (which translate to SortOp and ConstrainedSortOp respectively) in the IQT before the DbLimitExpression is visited.

Note that this fix interfered with changeset 3ceac74506d9 (Fix for OUTER APPLY regression with .NET 4.5) in that the original fix proposal was an additional rule being applied to the IQT that helped transform Outer Apply into a scalar subquery, but this Outer Apply was only the result of ITreeGenerator pushing the DbLimitExpression under the projection regardless of the type of operations contained by the it. This essentially means that the patterns we already had in place to cover the Outer Apply scenario were not being picked up because the IQT that we were producing did not match the patterns we look for until after we had applied other rules at least once: hence the reason the rule had to be applied again later down in the process.
There are two reasons that made me decide to revert the changes done by 3ceac74506d9:
1. Running the same rule again has performance considerations, as small as they might be they could affect larger queries.
2. The IQT that we are generating now doesn't require the re-application of any transformation rule. In effect, re-applying the transformation rule now would add unnecessary projections around the queries.
This change fixes the Outer Apply issue without affecting the resulting query, so no changes have been done to the test added by 3ceac74506d9, it is still passing.

Also note that there are many cases where we are still producing projections around the query that are unnecessary. This should be addressed in a different effort, and the issue is being tracked with workitem 1389.