Lazing on a Tuesday afternoon... (Fix extra lazy loading test)
This test would fail if it ran before other tests because the process of creating delegates to update collection instances dynamically would cause lazy loading to happen, which then altered the results. This only happens once because of caching, which is why the failure only happens when the test runs before other tests.