If you build the factory they will come (Fix for 1440: History context factory not resolved correctly)
The first issue was that the HistoryRepository allowed construction without specifying the factory explicitly. It would then do a simple lookup with a null key rather than doing the lookup with the provider invariant name and then doing the lookup with a null key only if this didn't resolve. The fix is to force HistoryRepository to always be passed a context factory and then to use the same mechanism we use for ContextKey to make sure the correct factory is passed.

The second issue is just a missing boilerplate method on DbConfiguration