Verbolicious (Adding verbs back to DbConfiguration)
After playing with the updated DbConfiguration DSL-style methods the general feeling is that they seem a bit strange, especially as they are not part of a fluent API. Therefore, this changes adds back the "Set" prefix for all except for interceptors, which gets "Add". Also, those methods that add services that are used by default (as opposed to only being used for some provider, context type, etc.) now also have "Default".