Your Public Image is Limited (Fix for 997: TypeLoadException thrown for a class nested in an internal class)
This was an issue with proxy creation of nested types. The proxy factory would check the visibility of the class in order to determine whether or not a new class in a different assembly is able to derive from it. Given a nested class like so:

internal class Outer
{
  public class Inner
  {
  }
}

Reflection returned the visibility of Inner as "PublicNested". We were therefore trying to derive from it to create the proxy. But even though Reflection says the class is public for all intents and purpose it is actually internal since it is nested inside an internal class. Therefore we cannot derive from it.

The fix is to correctly detect, despite what Reflection says, that types nested inside non-public types are not, practically speaking, public. We therefore do not attempt to create proxies for them.