<Everyone loves:xml="even though" xmlns:loves="to annoy everyone" />
Fix for 1499: Serialized SSDL contains auto generated namespace prefix for store extensions

When doing db reverse engineering the designer puts some attributes to the "http://schemas.microsoft.com/ado/2007/12/edm/EntityStoreSchemaGenerator" namespace. Since we did not write the namespace declaration on one of the ancestor elements the xml writer would write the namespace declaration on all elements that needed this namespace and - because we used it for attributes - it would also make up a safe namespace prefix similar to "p3" which made elements look like this:
<EntitySet Name="ChildEntities" EntityType="Self.ChildEntities" Schema="dbo" p3:Type="Tables" xmlns:p3="http://schemas.microsoft.com/ado/2007/12/edm/EntityStoreSchemaGenerator" />
This was regression for the designer where in the previous version we used to define the "store" namespace prefix for the namespace and used to write the namespace declaration on the Schema element which made elements look like this:
<EntitySet Name="ChildEntities" EntityType="Self.ChildEntities" Schema="dbo" store:Type="Tables"  />
The fix is to write the namespace declaration on the Schema element if needed (note currently we only check entity sets since this is the only place the designer is creating attributes in this namespace) which associates the prefix with the namespace uri and makes the xml writer use this prefix instead of inventing random prefixes on its own.