"They build us up and knocked us down" (Fix for 1524: Getting an ObjectContext from DbContextInfo-created DbContext can throw
When DbContextInfo is created with known provider information--i.e. a DbProviderInfo object--it uses this provider information to build the model. However, if an ObjectContext is then requested it is created using the normal connection discovery mechanism. An exception is then thrown if the supplied provider info does not match the provider information from the connection.

The fix is to use the provider information to override the connection just as we do when a connection is explicitly passed, but avoiding the case of doing this where a configuration from which to obtain the connection has been passed. The connection created is not a usable connection--it will throw if an attempt is made to access a database using it, but the point of this constructor is to avoid attempting to connect to the database anyway. Therefore we believe that while there is a chance of a break here it is minimal.