Less is the new black - Fix for codeplex workitems 199 and 486
Whenever a variable is referenced either by means of an .Include() or by the use of a navigation property, we pull in a ScanTable object into the query tree so that the variable can be used. The ScanTable will be created under an OuterApply which eventually might get promoted to a LeftOuterJoin operation. At this point, we will have one LeftOuterJoin operation per each reference to a member of an associated table.

Next, we would expect the join elimination rules to kick-in and eliminate the unnecessary join operations. However, in the particular shape of query described by workitems 199 and 486 the joins cannot be eliminated because they are not redundant: one is a LeftOuterJoin operation while the other one is an InnerJoin operation. But why do they differ?

It turns out that at an earlier stage the FilterOpRules class will attempt to promote left outer joins to inner joins given certain assumptions. This promotion is causing the otherwise identical join statements to differ, and the join elimination process to be unable to remove one or the other. The promotion of left outer joins to inner joins at this level is unnecessary and doesn't seem to bring any value. At a later stage in the process, after we have eliminated the redundant join statements, we have a separate and more robust process for promoting left outer joins into inner joins.

The fix to this bug is simply to eliminate the join promotion from FilterOpRules.