Not a generic customer (Fix for 1566: Initialization of custom collections is broken)
A collection navigation property type can be a custom type that implements ICollection<>. When materializing if the property is null, then we will attempt to create an instance of that type so that we can put entities in it. This is currently broken.

The problem was that the check for HashSet<> used to create the set with equality comparison was assuming that the collection type is generic. Custom collections may or may not be generic. The fix is to check whether or not the type is generic before attempting to get the type definition.