Nullbours, everybody needs good nullbours... (Fix for 1570: UpdateException being hidden by null ref exception in WrapUpdateException)
When WrapUpdateException was written is was called only in specific cases where the UpdateException being wrapped was always created with state entries. This is now no longer the case, which means an assert would fail in Debug code and a NullReferenceException would happen in Release code. Fixed to check for null and handle in the same way as an empty list of entries.