ForEach it may concern (Remove use of List.ForEach)
List.ForEach is not available in .NET for Windows Store apps. Replaced with existing Each extension method where available, and with foreach elsewhere. Each is not always available because it is an internal extension method, which is therefore not accessible in the functional tests. Redefinition is not trivial because then the unit tests would see both definitions. We discussed creating a differently named method, but decided it was better to just move to foreach.