Extreme Indentation Explained: Embrace Growth (Fix for 1603 EF6 rc1 Index was outside the bounds of the array Exception)
The problem was that the indented text writer which is used to format SQL was written to cache indents up to a point and then stop caching. But there was a bug in this code where the cache array would always be accessed even if the index was above the point where caching stopped. The fix is to cache always, growing the cache store as necessary.