Com S 228: Introduction to Data Structures (Fix for 1605: Increased memory usage EF5->EF6)
The reason for the increased memory usage in both EF5 and EF6 after SaveChanges until the collection of the context is that all the Added objects are converted to Unchanged objects in the state manager. Added objects do not have original values, while Unchanged objects do. The extra memory is used to track original values. The reason EF6 is using more memory here is due to a change made to improve the perf of original value tracking. Instead of using a List<StateManagerValue> we starting using a Dictionary<object, Dictionary<int, object>>.

The fix is to back out the change and start using List<StateManagerValue> again, where StateManagerValue is a value type. Some of the perf gains have been regained by reducing the number of linear scans and the number of times items are removed and then added back to the list. This has been done in a very restricted way at a local level to reduce risk. Further optimizations seem possible, but too risky at this time.

Unit tests were written for the current code and then moved over to the new code to help ensure that it is functionally the same.

In terms of memory usage, this change brings the increase in memory needed for tracking original values back down to EF5 levels.