Migrations: Fix for general scaffolding issues when upgrading from 5 due to model construction changes in 6.
The fix is to make the differ more robust by improving the way we match elements across the source and target models. Previously we would match tables and FKs based on their store name attributes, which would fail when the same element had different names in the source/target models. Elements can have different names because of:

1) Changes to the model construction algorithm - such as the determinism change.
2) Changes to the execution order of user ModelBuilder configuration statements. (We may want to address this post RTM)

The solution is start the match from the conceptual model - entity sets for tables and nav. props for FKs. We effectively trace a path from the conceptual elements, through mapping and down into the store model to determine which tables and FKs are the same element across the two models being diff'ed.

Results:

1) Table-splitting issue (#1576) - now generates an empty migration after upgrade from 5.
2) Many-to-many issue (http://www.davepaquette.com/archive/2013/08/25/entity-framework-6-rc1.aspx) - now generates a single rename table operation after upgrade from 5.
3) Although not a goal of this fix, the differ is now resilient to 2 above and also closer to something that can be used outside of CF.

Work Item: 1576