Unicornopoly: The PropertyInfo Trading Game (1667: Replace uses of Type.GetProperty, etc.)
The traditional methods for getting PropertyInfos from a Type are not available in .NET for Windows Store apps. The new extension methods should be used instead on .NET 4.5 and later. Where the new APIs are reasonable we start using and add support in the .NET 4 build. Where the new APIs are too cumbersome there are new extension methods tailored to the common things that we do.