That's soooo 2008... (1679: Model validation is failing for a valid model after changeset 082b451c17b)
The problem was that we were not restricting property lookup to declared properties only when scanning navigation properties in attribute-based mapping. This was not caught due to lack of coverage for attribute-based mapping. This has been added using dynamic assembly creation in the functional tests so that it does not impact all the POCO mapping we have. (POCO and attribute-based mapping cannot coexist in the same assembly.)