Copy of a... (First fix for 1454: Code First models still get built more than once in EF6)
The culprit here is EdmxWriter which is used to create an EDMX for use by Migrations and the differ. Previously it would always build the model again. Now if we are in the middle of doing DbContext initialization, then a model cached on the internal context is used instead. This cached model is released after initialization has finished so that we don't keep a potentially large model around forever.