Make PowerTools work again
Ever since EntityFramework.dll found its way into Visual Studio's assembly resolution path, the PowerTools have been broken. This fix updates the way we check if a type derives from DbContext. It ensures we use use the version of EF installed in the user's project and not the one VS finds by default. Some fix-up also had to be done to the user's config file in order to resolve the correct version. This work is part of #1165.