It's not you, it's me... (1678: AssemblyInformationalVersion is required by MigrationCodeGenerator)
The issue here was that CSharpMigrationCodeGenerator and VisualBasicMigrationCodeGenerator were using GetType() as a way of putting the EF assembly version into the generated code attribute. But if you are using a subclass of one of these generators, then you end up with the assembly version of the assembly containing the subclass. (This could be considered correct, but the original intent which still makes sense was to have the EF version here.) Further, if your assembly doesn't have an informational version attribute, then you get an exception. The fix is to always take the version from the EF assembly.