Dangerously Dangling Dexter (1142: Lazy Loading can return unexpected null)
The scenario here is one entity that contains two FK associations to the same target type. For example, A booking with two associated users. For the bug to repro the booking must be loaded and both associations must have the same FK value, but the user entity to which that FK refers must not be loaded. This results in an entry for the FK in the dangling FKs collection such that when the user with the matching PK is loaded we can do fixup for it. Now if only one of the FKs changes this results in the removal of the entry from the dangling FKs collection and the addition of a new entry for the changed FK. But this is wrong because the entry in the dangling FKs collection is still needed because of the other FK in the booking with the same value. This means that the fix is to not remove the entry from the dangling FKs collection if some other FK in the entity has the same FK value as the one being changed.