Tea for two; please don't throw... (546: Entity, with two references in base class to extending class)
The problem here was that we were trying to find a relationship by matching navigation properties, but in this scenario two relationships share the same navigation properties because they are lifted from the base type. Therefore we would get the "more than one element" exception. The fix is to also use the relationship name, which is now known at this point due to metadata unification.