It's abstract, but it's not art... (1289 & 1747: Mapping bug when abstract type is not explicitly added to the model)
Code First starts off by building EntityTypeMappings for a default TPH model. These mappings are not always valid because of at least two things: abstract types are excluded by default and some types in the TPH hierarchy must be changed to IsTypeOf mappings. Abstract types are excluded by default because they sometimes will not appear in the mappings. However, quite often they will appear, so we later attempt to add back abstract types if they are discovered in some other way--such as being configured explicitly. This means that later still when the entity mappings are fixed up for IsTypeOf changes and to remove unused fragments the code must take into account abstract types. So not mapping them by default was not much help and made it harder to find issues because, as in these bugs, the mappings for the abstract types were not added by any other path and so never got added.

The fix is to include abstract types by default and then let the later code remove them if not necessary when doing the mapping fix-up. This change revealed a couple of bugs in the fix-up code where abstract types were not being considered; these have been fixed. Also, we were using the lack of a type mapping as an indication of an abstract type with no mapped derived types in order to throw an early exception for this invalid mapping. That check is now moved to the point where we process all the entity types in the model.