DisorderlyConduct - fixing bug https://entityframework.codeplex.com/workitem/1735: Designer: Can not reverse engineer EDMX when FKs in different order than PKs (composite keys)
When validating the model we assumed that the order of primary key columns of the principal entity is the same as the order of the foreign key columns of the dependent entity. This is true for models built using Code First but does not have to be true for database first models where the order of columns is already defined in the database. Fixing validation to look the PK properties up by name based on the values specified in ReferentialConstraint for the prinicpal entity.