Remember My Name (1751: Unable to cast Anonymous Type from 'System.Linq.IQueryable' to 'System.Data.Entity.Core.Objects.ObjectQuery')
This was a regression from EF5 caused by the fix to 396: Handle closures containing DbSet/DbQuery members. However, this fix wasn't really the problem, it just exposed another existing bug that would previously only manifest in less common cases. This is because the fix for 396 used the type of the IQueryable as the type of the IQueryable, whereas before the type of the expression tree was used. In the common case (anonymous projection with no casting) for this scenario the type of the expression tree was IQueryable while the type of the IQueryable was actually ObjectQuery. Later on we made the assumption that the type would be IQueryable and not ObjectQuery. This was a false assumption as sometimes the expression tree type could also be ObjectQuery (see some tests) and so this would then fail. The fix is to treat ObjectQuery, IQueryable, and IOrderedQueryable the same as far as CQT generation is concerned instead of attempting to introduce an unnecessary cast, which then fails.