Adding the infrastructure to add an interceptor that can be used handle connection failures on transaction commit. The interceptor is resolved as a dependency of type Func<TransactionHandler> with the provider invariant name and database name as the key.
CommitFailureHandler is the implementation of a strategy that employs a TransactionContext to assign and store a Guid and Timestamp for each transaction in a seperate table in the same database as the target ObjectContext.
The TransactionContext is resolved as a dependency of type Func<DbConnection, TransactionContext> with the provider invariant name and database name as the key.

By default the initializer for TransactionContext will poll the database and add the missing table in case of a failure. This can be avoided by setting a different initializer for the TransactionContext-derived class used. TransactionContext. TransactionHandler.BuildDatabaseInitializationScript() can be used to get the T-SQL to perform the appropriate schema changes before executing the application.

Workitem: 1114