The Underground in November 1970 (765: Consider generalized mechanism for registering services in config file)
DbConfiguration has a Loaded event which is fired when the configuration has been loaded and is just about to be locked. This is very useful for third-party utilities who need to wrap or modify configuration as EF is initializing. However, it is not always easy for third-party code to find a place to run the code that will register the event. This change allows Loaded event handlers to be registered in the app.config or web.config, which is a place that third-party code can easily change, especially if installed by NuGet.