EF 6.0.2 - Fix for #1742 - EF6 regression: identity pattern not applied to key by convention in simple inheritance scenario
Problem: Changeset #9a111e568cea (fix for #1101) introduced a known breaking change where we would no longer introduce SGP.Identity for entities with self-refs. We thought this change would be low impact but it turns out to be higher than we anticipated as it affects some common self-ref, inheritance model patterns.
Solution: Revert the change. It was originally needed to fix #1101 but an alternative solution is to ensure that we always set fake keys when performing modification command tree generation.

Work Item: 1742