Fix for 1769: Initializer that queries a set can corrupt context internal state
It turns out that this only happens in limited cases and is not a regression from EF5. When we do initialization we create a temp ObjectContext that is used in the Seed method and elsewhere. After initialization this temp context is thrown away (to avoid issues like this) and the DbSet objects are reset such that any associated ObjectQuery is also thrown away since these are also associated with the temp context. Normally this means that any LINQ query done in the main context after initialization gets a new ObjectQuery and ObjectQueryProvider.

However, the ObjectQueryProvider can also be cached on the DbQueryProvider. This normally doesn't matter because the DbQuery is itself thrown away. But if the DbQuery is actually the DbSet, then it is not thrown away. This means that the cached ObjectQueryProvider (from the temp context) is retained. When this is then used it throws because the temp context has been disposed.

The fix is to not store the provider directly, but instead always obtain it from the query, which is itself reset when disposing the temp context.