Allow 'Code First' annotations to flow through the EDMX. This is required for the index attribute (item 57).
There was already code in the EDM serializer and deserializer to handle serialization of custom metadata annotations. The only annotations that are serialized are those that have a name with the format "A:B" where A is treated as an XML namespace and B as the XML attribute name. This change tests that this works and adds some infrastructure around it by serializing the ClrType attributes that Code First already uses.

In addition to this being a test case for annotation serialization it also paves the way to allow Code First-style o/c loading when using an EDMX, which would then avoid the naming collisions and nested type limitations that EDMX currently has. (Full support for this requires changes to the designer so that it annotates the EDMX with appropriate CLR type names.) An EDMX can flag its use of the Code First loaded by adding an annotation to the EDM container.

A new service has been added that allows for the serialization and deserialization of annotations such that annotations do not have to have just string values.